# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Chef:
    properties:
      chef_profile:
        $ref: '#/definitions/models.ChefProfile'
      chef_profile_id:
        type: integer
      cookie_expiration:
        description: nanoseconds
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.ChefProfile:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      fhs_certificate:
        description: Store file path (image/pdf)
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      preferred_name:
        type: string
      profile_picture:
        description: Store file path (.jpg, .png)
        type: string
      signature:
        description: Store file path or signature data
        type: string
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
  models.CreateChefProfileRequest:
    properties:
      address:
        type: string
      description:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      preferred_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
  models.CreateChefRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  models.UpdateChefProfileRequest:
    properties:
      address:
        type: string
      description:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      preferred_name:
        type: string
      verified:
        type: boolean
    type: object
  models.UpdateChefRequest:
    properties:
      email:
        type: string
      status:
        type: string
      username:
        minLength: 3
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Backend API for the Dished chef platform.
  title: Dished API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login as a chef
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new chef
      tags:
      - auth
  /chef-profiles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChefProfile'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all chef profiles
      tags:
      - chef-profiles
    post:
      consumes:
      - application/json
      parameters:
      - description: Profile payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateChefProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ChefProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a chef profile
      tags:
      - chef-profiles
  /chef-profiles/{id}:
    delete:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a chef profile
      tags:
      - chef-profiles
    get:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChefProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a chef profile by ID
      tags:
      - chef-profiles
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateChefProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChefProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a chef profile
      tags:
      - chef-profiles
  /chefs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Chef'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all chefs
      tags:
      - chefs
    post:
      consumes:
      - application/json
      parameters:
      - description: Create chef payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateChefRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Chef'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new chef
      tags:
      - chefs
  /chefs/{id}:
    delete:
      parameters:
      - description: Chef ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a chef
      tags:
      - chefs
    get:
      parameters:
      - description: Chef ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Chef'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a chef by ID
      tags:
      - chefs
    put:
      consumes:
      - application/json
      parameters:
      - description: Chef ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateChefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Chef'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a chef
      tags:
      - chefs
  /chefs/{id}/profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: Chef ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateChefProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a chef's profile
      tags:
      - chefs
  /chefs/usernames:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all chef usernames
      tags:
      - chefs
swagger: "2.0"